+++ boards/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 
 from .models import Board, Topic, Post
 from .forms import NewTopicForm
@@ -15,24 +16,30 @@ def board_topics(request, pk):
 
     return render(request, 'topics.html', {'board' : board})
 
+@login_required
 def new_topic(request, pk):
     board = get_object_or_404(Board, pk=pk)
-    user = User.objects.first() #Todo get logged in username
+
 
     if request.method == 'POST':
         form = NewTopicForm(request.POST)
         if form.is_valid():
             topic = form.save(commit=False)
             topic.board = board
-            topic.starter = user
+            topic.starter = request.user
             topic.save()
             post = Post.objects.create(
                 message=form.cleaned_data.get('message'),
                 topic = topic,
-                created_by = user
+                created_by = request.user
             )
             return redirect('board_topics', pk=board.pk)
     else:
         form = NewTopicForm()
 
     return render(request, 'new_topic.html', {'board' : board, 'form' : form})
+
+def topic_posts(request, pk, topic_pk):
+    topic = get_object_or_404(Topic, board__pk=pk, pk=topic_pk)
+
+    return render(request, 'topics_posts.html', {'topic'  : topic})
