+++ templates/topics.html
@@ -27,11 +27,11 @@
   <tbody>
     {% for topic in board.topics.all %}
       <tr>
-        <td>{{ topic.subject }}</td>
-        <td>{{ topic.starter.username }}</td>
-        <td>0</td>
-        <td>0</td>
-        <td>{{ topic.last_updated }}</td>
+        <td><a href="{% url 'topic_posts' board.pk topic.pk %}">{{ topic.subject }}</a></td>
+      <td>{{ topic.starter.username }}</td>
+      <td>0</td>
+      <td>0</td>
+      <td>{{ topic.last_updated }}</td>
       </tr>
     {% endfor %}
   </tbody>
