+++ templates/topics_posts.html
@@ -1,0 +1,7 @@
+{% extends 'base.html' %}
+
+{% load widget_tweaks %}
+
+{% block title %}
+  {{ topic.subject }}
+{% endblock %}
